// packages/tron/src/gql/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K];
};
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]?: Maybe<T[SubKey]>;
};
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]: Maybe<T[SubKey]>;
};
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  AssetV0Args: any;
  /**
   * A datetime with timezone offset.
   *
   * The input is a string in RFC3339 format, e.g. "2022-01-12T04:00:19.12345Z"
   * or "2022-01-12T04:00:19+03:00". The output is also a string in RFC3339
   * format, but it is always normalized to the UTC (Z) offset, e.g.
   * "2022-01-12T04:00:19.12345Z".
   */
  DateTime: any;
  /** Decimal (fixed-point) */
  Decimal: any;
  IntegerString: any;
  /** A scalar that can represent any JSON value. */
  JSON: any;
  /**
   * A local datetime without timezone offset.
   *
   * The input/output is a string in ISO 8601 format without timezone, including
   * subseconds. E.g. "2022-01-12T07:30:19.12345".
   */
  LocalDateTime: any;
};
// ...
export type AssetAmountType = {
  __typename?: 'AssetAmountType';
  /** Currency allTimeHigh */
  allTimeHigh?: Maybe<Scalars['String']>;
  /** Currency allTimeLow */
  allTimeLow?: Maybe<Scalars['String']>;
  /** Currency price amount */
  amount: Scalars['String'];
  /** Currency dailyHigh */
  dailyHigh?: Maybe<Scalars['String']>;
  /** Currency dailyLow */
  dailyLow?: Maybe<Scalars['String']>;
  /** 24 Hour Trading Volume */
  dailyTradingVolume?: Maybe<Scalars['String']>;
  /** Currency dayPriceChange */
  dayPriceChange?: Maybe<Scalars['String']>;
  /** Currency fdv */
  fdv?: Maybe<Scalars['String']>;
  /** Currency marketCapRank */
  marketCapRank?: Maybe<Scalars['String']>;
  /** Currency monthPriceChange */
  monthPriceChange?: Maybe<Scalars['String']>;
  /** The scaling factor is needed to convert the amount of payment to the currency */
  scalingFactor: Scalars['Float'];
  /** Current sparkline */
  sparkline: Array<Scalars['Float']>;
  /** Current sparkline image url */
  sparklineImageUrl?: Maybe<Scalars['String']>;
  /** Updated on */
  updatedOn: Scalars['Float'];
  /** Currency weekPriceChange */
  weekPriceChange?: Maybe<Scalars['String']>;
  /** Currency yearPriceChange */
  yearPriceChange?: Maybe<Scalars['String']>;
};
// ...
export enum AssetInternalType {
  CRYPTOCURRENCY = 'CRYPTOCURRENCY',
  LP_TOKEN = 'LP_TOKEN',
  TOKEN = 'TOKEN',
}
// ...
export type CryptoAsset = {
  __typename?: 'CryptoAsset';
  /** Identify asset as a shitcoin, stablecoin, lp token, lst token or trending token. */
  categories: Array<TokenCategory>;
  /** supported list of chain are in [`crate::chain::Chain`] enum */
  chain?: Maybe<Scalars['String']>;
  /** ID of token (contract address in most chain) */
  contract?: Maybe<Scalars['String']>;
  /** Number of decimals for current asset */
  decimals?: Maybe<Scalars['Float']>;
  /** Unique asset identifier */
  id?: Maybe<Scalars['ID']>;
  /** Asset image */
  image?: Maybe<Scalars['String']>;
  /** Known name that identifies token */
  name?: Maybe<Scalars['String']>;
  price?: Maybe<AssetAmountType>;
  /** The symbol that identifies token */
  symbol?: Maybe<Scalars['String']>;
  type?: Maybe<AssetInternalType>;
};
// ...
export enum TokenCategory {
  LP_TOKEN = 'LP_TOKEN',
  LST_TOKEN = 'LST_TOKEN',
  SHITCOIN = 'SHITCOIN',
  STABLECOIN = 'STABLECOIN',
  TRENDING_TOKEN = 'TRENDING_TOKEN',
}